requirements.txt
numpy==1.26.4
pyaudio==0.2.14
sounddevice==0.4.6
soundfile==0.12.1
PyDrive==1.3.1
PyYAML==6.0.1
